+++ public/scripts/checker/checker-events.js
@@ -1,0 +1,11 @@
+/*
+Name:
+    checker-events.js
+Purpose:
+    This file is designed for the purpose of handling all of the event setup for the balance checker.
+    Many of the methods in this file are designed to only be called once. This is to prevent multiple event listeners from being added to the same element.
+    Please see the method mapper for more information.
+*/
+
+
+
